clean repo boiler plate
