+++ pages/index.jsx
@@ -1,0 +1,25 @@
+import Head from 'next/head'
+import Image from 'next/image'
+import styles from '../styles/Home.module.css'
+
+const Home = () => {
+  return (
+    <div className={styles.container}>
+      <Head>
+        <title>Mazen Bahgat Portfolio</title>
+        <meta name="Mazen Bahgat's Development Portfolio" content="Software Development Skills and Experiences" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <div className={styles.main}>
+        
+      </div>
+
+      <footer className={styles.footer}>
+        
+      </footer>
+    </div>
+  )
+}
+
+export default Home;
